+++ android/src/main/kotlin/de/mintware/barcode_scan/BarcodeScannerActivity.kt
@@ -109,7 +109,11 @@ class BarcodeScannerActivity : Activity(), ZXingScannerView.ResultHandler {
         super.onResume()
         setupScannerView()
         scannerView?.setResultHandler(this)
-        scannerView?.startCamera()
+        if (config.useCamera > -1) {
+            scannerView?.startCamera(config.useCamera)
+        } else {
+            scannerView?.startCamera()
+        }
     }
     // endregion